bugfix: 规范目录接口契约测试分层
